refactor(GameEndService): UserBallService의 isAnswer 기능 GameEndService 클래스로 분리
- UserBallService 는 userball 관리하는 서비스라 정답을 판단하는 기능하고 어울리지 않음
- GameEndService 생성에 따른 froncoller 의존성 및 관련 로직 수정 , GameConfig 도 수정
